fix(trait): synthetic Kit
While working on #5378 I realized the concept of external Kit is used widely also for the concept of what is a Synthetic Kit.

An external Kit is a kit coming from a build from the operator, whilst a synthetic Kit is not coming from an operator.

(cherry picked from commit apache/camel-k@b895cfc90)